packaging: fix (workaround) socket-based activation to run as app.
This patch also takes out all the systemd-related conditionality
from the spec file. It seems that conditional inclusion in the
%files section stopped working at some point in time. The exact
same conditionals work in all other sections but not in %files.
So while one can still configure and build the package based on
various options, the %files section can't be adjusted dynamically
with a simple %if any more because it always evaluates to false.
I guess the only option is to generate a file list and use that
instead. Not very convenient...

Change-Id: I6186209921501b235ad25b77e5db8edd81a3b2b6